Add HFL Example and Fix Some Issues
example3.ml implements a powerset lattice and corresponding HFL formula. The elements of the lattice are
integer sets, which makes it a straightforward implementation. one could consider a simpler datatype as
the lattice is restricted to this specific example.

due to the datatype t in example3.ml it was necessary to extend the signature of Lattice in HOFPIteration.ml
by an equality check  function. this is used to compare two elements of a specific lattice instance.

extended Makefile with example3.ml

fixed minor bug in example1.ml regarding nun function